fix type check
